requirements.txt
flask==3.0.0
pillow==10.1.0
openai==1.54.0
python-dotenv==1.0.0
werkzeug==3.0.1
